descriptor_utils: Simplify Reader/Writer::new()
There is no reason to go through the `GuestMemoryRegion` to get a slice
in guest memory, we can just use `GuestMemory::get_slice()` directly.

This is not least in preparation for a vm-memory update that makes
`GuestMemory` potentially represent virtual memory, in which case we
will no longer be able to easily translate memory addresses into
`GuestMemoryRegion`s.

Another change that will come with that is that `get_slice()` will be
removed in favor of an iterating `get_slices()`.  This would require us
replacing `.map()` with `.flat_map()`, but with us generating errors on
both iterator levels (in the `chain.readable()`/`chain.writable()`
iterator, and in the `.get_slices()` iterator), this is quite hard to
represent in this functional style.

So just use a `for .. in` loop with `.push_back()` into the `VecDeque`.
Given that none of our iterators has a `.size_hint()` implementation, it
doesn’t make a difference in performance anyway.

And while at it, we may as well already use `get_slices()` rather than
`get_slice()`.  (And similarly, push the slices in a `for .. in` loop
because there is no `Vec::try_extend()`.)

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>